+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+[1.0.4] - 2025-01-07
+
+Added
+* Added a button to select files or directories for YAML processing
+* Integrated IntelliJ IDEA's file chooser for selecting files or directories
+
+Changed
+* Reorganized buttons in the UI for better usability
+* Renamed buttons to more appropriate names
+
 [1.0.3] - 2024-12-26
 
 Added